Removed mappings which changed vim defaults
I'd switched the meanings of ` and ' because I mostly want the
functionality of ` however on a Swedish keyboard if you hit `a you get à
and the jump fails. However I don't use this enough to justify changing
it from the default. All I need to do to make the default work is stay
away from vowels. `m for example works great.